sdl fixes

+++ src/uws_client.c
@@ -831,7 +831,7 @@ static void on_underlying_io_open_complete(void* context, IO_OPEN_RESULT open_re
                     }
                     else
                     {
-                        upgrade_request = (char*)malloc(upgrade_request_length + 1);
+                        upgrade_request = (char*)malloc((size_t)upgrade_request_length + 1);
                         if (upgrade_request == NULL)
                         {
                             /* Codes_SRS_UWS_CLIENT_01_406: [ If not enough memory can be allocated to construct the WebSocket upgrade request, uws shall report that the open failed by calling the on_ws_open_complete callback passed to uws_client_open_async with WS_OPEN_ERROR_NOT_ENOUGH_MEMORY. ]*/
